requirements.txt
kivy==2.2.1
kivymd==1.1.1
Pillow==10.0.0
kivy_garden.mapview==1.0.6 